pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
replit = "^3.2.4"
Flask = "^2.2.0"
urllib3 = "^1.26.12"
python-dotenv = "^0.21.0"
websocket-server = "^0.6.4"
websocket = "^0.2.1"
websockets = "^10.4"
tqdm = "^4.64.1"
Flask-SocketIO = "^5.3.2"
flask-sqlalchemy = "^3.0.2"
SQLAlchemy = "^1.4.45"
progress = "^1.6"
cryptography = "^38.0.4"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
replit-python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.9"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"